for RPI

+++ examples/gles2/fragment_tile.glsl
@@ -1,22 +1,22 @@
 uniform sampler2D texture;
 
-varying mediump vec2 texcoord;
-uniform mediump vec2 texSize;
-uniform mediump float pixelscale;
-uniform mediump float prio;
-uniform mediump vec2 tilesize;
+varying highp vec2 texcoord;
+uniform highp vec2 texSize;
+uniform highp float pixelscale;
+uniform highp float prio;
+uniform highp vec2 tilesize;
 
 void main(void) {
-  mediump vec4 tile = texture2D(texture, texcoord / texSize);
+  highp vec4 tile = texture2D(texture, texcoord / texSize);
   tile *= 255.0;
-  // mediump float prioFlag = tile.z;
-  mediump float prioSet = mod(prio, 2.0);
-  mediump float mult = prioSet == tile.z ? 1.0 : 0.0; 
+  // highp float prioFlag = tile.z;
+  highp float prioSet = mod(prio, 2.0);
+  highp float mult = prioSet == tile.z ? 1.0 : 0.0; 
   tile *= mult;
 
-  mediump vec2 oneTile = (texSize / tilesize);
+  highp vec2 oneTile = (texSize / tilesize);
   
-  mediump vec2 offset = fract(texcoord);
+  highp vec2 offset = fract(texcoord);
 
   gl_FragColor = texture2D(texture, (tile.xy + offset) / oneTile);
 }
